+++ maverics/src/main/kotlin/com/example/services/AddPlatformCharge.kt
@@ -3,6 +3,6 @@ package com.example.services
 import com.example.constants.totalPlatformFees
 import java.math.BigInteger
 
-fun addPlatformCharge(platformCharge: Long){
-    totalPlatformFees += BigInteger.valueOf(platformCharge)
+fun addPlatformCharge(platformCharge: BigInteger) {
+    totalPlatformFees += platformCharge
 }
