+++ maverics/src/main/kotlin/com/example/services/GetWalletInfo.kt
@@ -1,10 +1,11 @@
 package com.example.services
 
 import com.example.controller.walletList
+import java.math.BigInteger
 
 
-fun getWalletInfo(username: String): MutableMap<String, Long> {
-    val walletInfo = mutableMapOf<String, Long>()
+fun getWalletInfo(username: String): MutableMap<String, BigInteger> {
+    val walletInfo = mutableMapOf<String, BigInteger>()
     val userWallet = walletList[username]
     walletInfo["free"] = userWallet!!.freeAmount
     walletInfo["locked"] = userWallet.lockedAmount
